RELATED: RAIL-655 Split VO's for UI SDK and saving
