+++ src/internal/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Internal
+ *
+ * Internal module contains interfaces interfaces used internally in GoodData apps.
+ */
+
+export * from './VisualizationObject';
